+++ integration-tests/cloudwatch-logwriter-integration-tests/src/test/java/com/kdgregory/logging/aws/CloudWatchLogWriterIntegrationTest.java
@@ -14,6 +14,10 @@
 
 package com.kdgregory.logging.aws;
 
+import java.lang.Thread.UncaughtExceptionHandler;
+import java.util.ArrayList;
+import java.util.List;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -36,6 +40,7 @@ import com.kdgregory.logging.aws.cloudwatch.CloudWatchWriterConfig;
 import com.kdgregory.logging.aws.cloudwatch.CloudWatchWriterFactory;
 import com.kdgregory.logging.aws.cloudwatch.CloudWatchWriterStatistics;
 import com.kdgregory.logging.common.LogMessage;
+import com.kdgregory.logging.common.LogWriter;
 import com.kdgregory.logging.common.factories.DefaultThreadFactory;
 import com.kdgregory.logging.common.util.DiscardAction;
 import com.kdgregory.logging.testhelpers.CloudWatchTestHelper;
@@ -45,7 +50,7 @@ import com.kdgregory.logging.testhelpers.TestableInternalLogger;
 
 public class CloudWatchLogWriterIntegrationTest
 {
-    private final static String LOGGROUP_NAME = "CloudWatchLogWriterIntegrationTest";
+    private final static String BASE_LOGGROUP_NAME = "CloudWatchLogWriterIntegrationTest";
 
     // single "helper" client that's shared by all tests
     private static AWSLogsClient helperClient;
@@ -60,43 +65,86 @@ public class CloudWatchLogWriterIntegrationTest
     private Logger localLogger = LoggerFactory.getLogger(getClass());
 
     // these are all assigned by init()
+    private String logGroupName;
+    private String logStreamName; // default, tests may ignore
     private CloudWatchTestHelper testHelper;
     private CloudWatchWriterStatistics stats;
     private TestableInternalLogger internalLogger;
-    private CloudWatchWriterConfig config;
-    private CloudWatchWriterFactory factory;
-    private CloudWatchLogWriter writer;
+    private CloudWatchWriterFactory writerFactory;
+    private DefaultThreadFactory threadFactory;
+
+    // tests that create multiple writers/threads will find them here
+    private List<Thread> writerThreads = new ArrayList<Thread>();
+    private List<CloudWatchLogWriter> writers = new ArrayList<CloudWatchLogWriter>();
 
 //----------------------------------------------------------------------------
 //  Helpers
 //----------------------------------------------------------------------------
 
-    private void init(String testName, Integer retentionPeriod, AWSLogs client, String factoryMethod, String region, String endpoint)
+    /**
+     *  Called at the beginning of most tests to create the test helper and
+     *  writer, and delete any previous log group.
+     */
+    private void init(String testName, AWSLogs client)
     throws Exception
     {
         MDC.put("testName", testName);
         localLogger.info("starting");
 
-        testHelper = new CloudWatchTestHelper(client, LOGGROUP_NAME);
+        logGroupName = getClass().getSimpleName() + "-" + testName;
+        logStreamName = testName;
 
+        testHelper = new CloudWatchTestHelper(client, logGroupName);
         testHelper.deleteLogGroupIfExists();
 
-        config = new CloudWatchWriterConfig(LOGGROUP_NAME, testName, retentionPeriod, 250, 10000, DiscardAction.oldest, factoryMethod, region, endpoint);
         stats = new CloudWatchWriterStatistics();
         internalLogger = new TestableInternalLogger();
-        factory = new CloudWatchWriterFactory();
-        writer = (CloudWatchLogWriter)factory.newLogWriter(config, stats, internalLogger);
+        writerFactory = new CloudWatchWriterFactory();
 
-        new DefaultThreadFactory("test").startLoggingThread(writer, false, null);
+        threadFactory = new DefaultThreadFactory("test")
+        {
+            @Override
+            protected Thread createThread(LogWriter logWriter, UncaughtExceptionHandler exceptionHandler)
+            {
+                Thread thread = super.createThread(logWriter, exceptionHandler);
+                writerThreads.add(thread);
+                return thread;
+            }
+        };
+    }
+
+
+    /**
+     *  Returns a default writer config, which the test can then modify.
+     */
+    private CloudWatchWriterConfig defaultConfig()
+    {
+        return new CloudWatchWriterConfig(logGroupName, logStreamName, null, false, 250, 10000, DiscardAction.oldest, null, null, null);
+    }
+
+
+    /**
+     *  Creates the log writer and its execution thread, and starts it running.
+     */
+    private CloudWatchLogWriter createWriter(CloudWatchWriterConfig config)
+    throws Exception
+    {
+        CloudWatchLogWriter writer = (CloudWatchLogWriter)writerFactory.newLogWriter(config, stats, internalLogger);
+        writers.add(writer);
+        threadFactory.startLoggingThread(writer, false, null);
+        return writer;
     }
 
 
     private class MessageWriter
     extends com.kdgregory.logging.testhelpers.MessageWriter
     {
-        public MessageWriter(int numMessages)
+        private CloudWatchLogWriter writer;
+
+        public MessageWriter(CloudWatchLogWriter writer, int numMessages)
         {
             super(numMessages);
+            this.writer = writer;
         }
 
         @Override
@@ -135,7 +183,7 @@ public class CloudWatchLogWriterIntegrationTest
     @After
     public void tearDown()
     {
-        if (writer != null)
+        for (CloudWatchLogWriter writer : writers)
         {
             writer.stop();
         }
@@ -164,12 +212,15 @@ public class CloudWatchLogWriterIntegrationTest
     {
         final int numMessages = 1001;
 
-        init("smoketest", null, helperClient, null, null, null);
+        init("smoketest", helperClient);
 
-        new MessageWriter(numMessages).run();
+        CloudWatchWriterConfig config = defaultConfig();
+        CloudWatchLogWriter writer = createWriter(config);
+
+        new MessageWriter(writer, numMessages).run();
 
         CommonTestHelper.waitUntilMessagesSent(stats, numMessages, 30000);
-        testHelper.assertMessages("smoketest", numMessages);
+        testHelper.assertMessages(logStreamName, numMessages);
 
         assertNull("static factory method not called", factoryClient);
 
@@ -182,12 +233,16 @@ public class CloudWatchLogWriterIntegrationTest
     {
         final int numMessages = 1001;
 
-        init("testFactoryMethod", null, helperClient, getClass().getName() + ".staticClientFactory", null, null);
+        init("testFactoryMethod", helperClient);
 
-        new MessageWriter(numMessages).run();
+        CloudWatchWriterConfig config = defaultConfig();
+        config.clientFactoryMethod = getClass().getName() + ".staticClientFactory";
+        CloudWatchLogWriter writer = createWriter(config);
+
+        new MessageWriter(writer, numMessages).run();
 
         CommonTestHelper.waitUntilMessagesSent(stats, numMessages, 30000);
-        testHelper.assertMessages("testFactoryMethod", numMessages);
+        testHelper.assertMessages(logStreamName, numMessages);
 
         assertNotNull("factory method was called", factoryClient);
         assertSame("factory-created client used by writer", factoryClient, ClassUtil.getFieldValue(writer, "client", AWSLogs.class));
@@ -202,15 +257,19 @@ public class CloudWatchLogWriterIntegrationTest
         // default region for constructor is always us-east-1
         altClient = new AWSLogsClient().withRegion(Regions.US_WEST_1);
 
-        init("testAlternateRegion", null, altClient, null, "us-west-1", null);
+        init("testAlternateRegion", altClient);
 
-        new MessageWriter(numMessages).run();
+        CloudWatchWriterConfig config = defaultConfig();
+        config.clientRegion = "us-west-1";
+        CloudWatchLogWriter writer = createWriter(config);
+
+        new MessageWriter(writer, numMessages).run();
 
         CommonTestHelper.waitUntilMessagesSent(stats, numMessages, 30000);
-        testHelper.assertMessages("testAlternateRegion", numMessages);
+        testHelper.assertMessages(logStreamName, numMessages);
 
         assertFalse("stream does not exist in default region",
-                    new CloudWatchTestHelper(helperClient, LOGGROUP_NAME).isLogStreamAvailable("testAlternateRegion"));
+                    new CloudWatchTestHelper(helperClient, BASE_LOGGROUP_NAME).isLogStreamAvailable("testAlternateRegion"));
     }
 
 
@@ -224,27 +283,71 @@ public class CloudWatchLogWriterIntegrationTest
         //         if that is your default, then the test will fail
         altClient = new AWSLogsClient().withEndpoint("logs.us-east-2.amazonaws.com");
 
-        init("testAlternateEndpoint", null, altClient, null, null, "logs.us-east-2.amazonaws.com");
+        init("testAlternateEndpoint", altClient);
 
-        new MessageWriter(numMessages).run();
+        CloudWatchWriterConfig config = defaultConfig();
+        config.clientEndpoint = "logs.us-east-2.amazonaws.com";
+        CloudWatchLogWriter writer = createWriter(config);
+
+        new MessageWriter(writer, numMessages).run();
 
         CommonTestHelper.waitUntilMessagesSent(stats, numMessages, 30000);
-        testHelper.assertMessages("testAlternateEndpoint", numMessages);
+        testHelper.assertMessages(logStreamName, numMessages);
 
         assertFalse("stream does not exist in default region",
-                    new CloudWatchTestHelper(helperClient, LOGGROUP_NAME).isLogStreamAvailable("testAlternateEndpoint"));
+                    new CloudWatchTestHelper(helperClient, BASE_LOGGROUP_NAME).isLogStreamAvailable("testAlternateEndpoint"));
     }
 
 
     @Test
     public void testRetentionPeriod() throws Exception
     {
-        init("testRetentionPeriod", 3, helperClient, null, null, null);
+        init("testRetentionPeriod", helperClient);
+
+        CloudWatchWriterConfig config = defaultConfig();
+        config.retentionPeriod = 3;
+        CloudWatchLogWriter writer = createWriter(config);
 
         // will write a single message so that we have something to wait for
-        new MessageWriter(1).run();
+        new MessageWriter(writer, 1).run();
         CommonTestHelper.waitUntilMessagesSent(stats, 1, 30000);
 
         assertEquals("retention period", 3, testHelper.describeLogGroup().getRetentionInDays().intValue());
+    }
+
+
+    @Test
+    public void testDedicatedWriter() throws Exception
+    {
+        final int numWriters = 10;
+        final int numReps = 50;
+
+        init("testDedicatedWriter", helperClient);
+
+        for (int ii = 0 ; ii < numWriters ; ii++)
+        {
+            localLogger.debug("creating writer {}", ii);
+            CloudWatchWriterConfig config = defaultConfig();
+            config.logStreamName = "testDedicatedWriter-" + ii;
+            config.dedicatedWriter = true;
+            createWriter(config);
+        }
+
+        for (int ii = 0 ; ii < numReps ; ii++)
+        {
+            localLogger.debug("writing message {}", ii);
+            LogMessage message = new LogMessage(System.currentTimeMillis(), String.format("message %d", ii));
+            for (CloudWatchLogWriter w : writers)
+            {
+                // the sleep is intended to ensure that the messages will end up
+                // in separate batches; in practice, with the background threads
+                // all runnable, it won't come back for three seconds or more
+                w.addMessage(message);
+                Thread.sleep(100);
+            }
+        }
+
+        CommonTestHelper.waitUntilMessagesSent(stats, numWriters * numReps, 30000);
+        internalLogger.assertInternalErrorLog();
     }
 }